requirements.txt
discord.py
aiohttp
kaggle